Change wclear → werase :zap: :building_construction:
Using wclear forces the entire screen to be redrawn. From the Ncurses
documentation:

The erase and werase routines copy blanks to every position in the
window, clearing the screen. The clear and wclear routines are like
erase and werase, but they also call clearok, so that the screen is
cleared completely on the next call to wrefresh for that window and
repainted from scratch.

Architectural changes were made since the draw* methods no longer need
to be called separately.

Fixes #7.