Only render lines which are on screen :zap:
This improves performance when editing long files.

Scrolling when the cursor goes off-screen is difficult because the
screen size is only available in the rendering code, which cannot edit
state to set a new offset. Brick has a separate rendering state for
this, but it is internal, so I chose to simply lock the cursor in
the centre of the screen instead. Therefore implementation of #23 will
most likely be moving the cursor vertically by a predefined distance.